PowerShell script for Copy Database Backup File
In this script, we copy the database backup files from the database server to the backup servers. The copy was validated for not copying the file if it already exists in backup servers. We made the log also for documentation or debugging purposes. You just need to set the task scheduler on windows to run this task based on your needs.